refactor: 교육에 해당하는 결승 진출 조회 로직 API 엔드포인트 변경
- quiz -> education
